+++ BankAccountDDD/Account.DDD/Events/WireTransferred.cs
@@ -8,9 +8,9 @@ namespace Account.DDD.Events
     public class WireTransferred : Event
     {
         public readonly Guid AccountId;
-        public readonly double TransferredFunds;
+        public readonly decimal TransferredFunds;
 
-        public WireTransferred(Guid accountId, double transferredFunds)
+        public WireTransferred(Guid accountId, decimal transferredFunds)
         {
             AccountId = accountId;
             TransferredFunds = transferredFunds;
